+++ integreat_cms/cms/views/pages/page_actions.py
@@ -568,31 +568,31 @@ def grant_page_permission_ajax(request, region_slug):
             # check, if the user already has this permission
             if user.has_perm("cms.change_page_object", page):
                 message = _(
-                    "Information: The user {user} has this permission already."
-                ).format(user=user.username)
+                    'Information: The user "{}" has this permission already.'
+                ).format(user.full_user_name)
                 level_tag = "info"
             else:
                 # else grant the permission by adding the user to the editors of the page
                 page.editors.add(user)
                 page.save()
-                message = _("Success: The user {user} can now edit this page.").format(
-                    user=user.username
+                message = _('Success: The user "{}" can now edit this page.').format(
+                    user.full_user_name
                 )
                 level_tag = "success"
         elif permission == "publish":
             # check, if the user already has this permission
             if user.has_perm("cms.publish_page_object", page):
                 message = _(
-                    "Information: The user {user} has this permission already."
-                ).format(user=user.username)
+                    'Information: The user "{}" has this permission already.'
+                ).format(user.full_user_name)
                 level_tag = "info"
             else:
                 # else grant the permission by adding the user to the publishers of the page
                 page.publishers.add(user)
                 page.save()
-                message = _(
-                    "Success: The user {user} can now publish this page."
-                ).format(user=user.username)
+                message = _('Success: The user "{}" can now publish this page.').format(
+                    user.full_user_name
+                )
                 level_tag = "success"
         else:
             logger.warning("Error: The permission %r is not supported", permission)
@@ -680,14 +680,14 @@ def revoke_page_permission_ajax(request, region_slug):
             # check, if the user has this permission anyway
             if user.has_perm("cms.change_page_object", page):
                 message = _(
-                    "Information: The user {user} has been removed from the editors of this page, "
+                    'Information: The user "{}" has been removed from the editors of this page, '
                     "but has the implicit permission to edit this page anyway."
-                ).format(user=user.username)
+                ).format(user.full_user_name)
                 level_tag = "info"
             else:
                 message = _(
-                    "Success: The user {user} cannot edit this page anymore."
-                ).format(user=user.username)
+                    'Success: The user "{}" cannot edit this page anymore.'
+                ).format(user.full_user_name)
                 level_tag = "success"
         elif permission == "publish":
             if user in page.publishers.all():
@@ -697,14 +697,14 @@ def revoke_page_permission_ajax(request, region_slug):
             # check, if the user already has this permission
             if user.has_perm("cms.publish_page_object", page):
                 message = _(
-                    "Information: The user {user} has been removed from the publishers of this page, "
+                    'Information: The user "{}" has been removed from the publishers of this page, '
                     "but has the implicit permission to publish this page anyway."
-                ).format(user=user.username)
+                ).format(user.full_user_name)
                 level_tag = "info"
             else:
                 message = _(
-                    "Success: The user {user} cannot publish this page anymore."
-                ).format(user=user.username)
+                    'Success: The user "{}" cannot publish this page anymore.'
+                ).format(user.full_user_name)
                 level_tag = "success"
         else:
             logger.warning("Error: The permission %r is not supported", permission)